ecom_backend/config/wsgi.py
"""
WSGI config for ERP E-Commerce project.
"""
import os

# Auto-configure PyMySQL as MySQLdb if mysqlclient is not installed
try:
    import MySQLdb  # noqa: F401
except ImportError:
    try:
        import pymysql
        pymysql.install_as_MySQLdb()
    except ImportError:
        pass

from django.core.wsgi import get_wsgi_application

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'config.settings.production')

application = get_wsgi_application()
