ecom_backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0

# Authentication
djangorestframework-simplejwt==5.3.1

# Database - MySQL
mysqlclient==2.2.1
# Alternative (pure Python, no compilation): PyMySQL==1.1.0

# CORS
django-cors-headers==4.3.1

# Filtering & Pagination
django-filter==23.5

# Environment Variables
python-dotenv==1.0.0

# Image Processing
Pillow==10.2.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
faker==22.0.0

# Code Quality
flake8==7.0.0
black==23.12.1
isort==5.13.2

# Documentation
drf-spectacular==0.27.0

# Utilities
python-dateutil==2.8.2

# For development
django-extensions==3.2.3
